pyproject.toml
[project]
name = "pano-labeler"
version = "1.0.0"
description = "Web-based tool for annotating panoramic images with spherical bounding boxes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pillow==10.1.0",
    "numpy==1.26.2",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "pyyaml==6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]


[tool.uv]
dev-dependencies = []
